% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{correct_z}
\alias{correct_z}
\title{Select component corresponding to coverslip}
\usage{
correct_z(
  data,
  component = 1,
  coverslip = TRUE,
  median.filter.size = 2,
  threshold = 0.2
)
}
\arguments{
\item{data}{A raw sims data object}

\item{component}{Principal component corresponding to the coverslip}

\item{coverslip}{Logical: Is there a coverslip}

\item{median.filter.size}{Size of median kernel for filtering noise}

\item{threshold}{Threshold to identify coverslip signal}
}
\value{
A scaled data matrix
}
\description{
Select component corresponding to coverslip
}
