% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{open_raw_txt}
\alias{open_raw_txt}
\title{Function to load and pre-process ToF-SIMS images in .txt format}
\usage{
open_raw_txt()
}
\value{
A raw.sims.data object
}
\description{
Function to load and pre-process ToF-SIMS images in .txt format
}
