% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ppca}
\alias{ppca}
\title{Perform PCA with poisson scaling}
\usage{
ppca(data)
}
\arguments{
\item{data}{Raw sims data}
}
\value{
A scaled sims data object
}
\description{
Perform PCA with poisson scaling
}
